refactor(PurchaseAmount): 구매 금액 단위를 `int`에서 `long`으로 변경

+++ src/main/java/lotto/domain/ProfitCalculator.java
@@ -12,7 +12,7 @@ public class ProfitCalculator {
         List<LottoRank> ranks = LottoResultRepository.findLottoRankResults();
 
         long profitAmount = calculateTotalProfitAmount(ranks);
-        int purchaseAmount = PurchaseAmountRepository.findAmount();
+        long purchaseAmount = PurchaseAmountRepository.findAmount();
 
         return calculateTotalProfitRate(profitAmount, purchaseAmount);
     }
